Create README.md
In the concurrent world, the situation of this Covid-19 pandemic has been very disastrous and for the purpose of  reducing  the transmission of SARS virus, it has been made mandatory by many governments to wear a facial mask. This project basically uses machine learning techniques to recognize whether a person is wearing a face mask or not. This project can be used so as to detect whether a person is wearing the face mask and it can clearly differentiate between a person with facial mask covering or not. Many machine learning  Frameworks like Open CV, Tensor flow and Keras, Mobile net are used in this project.The model is quite accurate as the accuracy which was procured after proper implementation came out to be more than 90\%.In this paper, we have elaborated all the implementation of the code along with the accuracy, precision of the model to recognize facial covering.